backend/requirements.txt
fastapi==0.115.0
uvicorn==0.30.0
pydantic==2.9.0
python-multipart==0.0.9
python-jose==3.3.0
passlib==1.7.4
bcrypt==4.2.0
thefuzz==0.22.1
python-Levenshtein==0.25.1
supabase
python-dotenv
